實作 第一集買了一本，第二集也買了一本，價格應為100*2*0.95=190

+++ Hank_TDD_Day2Homework/Services/IShoppingCartService.cs
@@ -4,6 +4,6 @@
     {
         void OrderProduct(T product);
 
-        int Bill();
+        double Bill();
     }
 }
